Move `slotToUTCTime` conversion to `Cardano.Wallet.Deposit.Time`
